requirements.txt
html5lib==1.0b3
PyYAML==3.11
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
contextlib2==0.4.0
flake8==2.3.0
jira==1.0.3
mccabe==0.3
mock==1.0.1
nose==1.3.4
pep8==1.6.2
pyflakes==0.8.1
python-dateutil==2.2
requests==2.5.2
six==1.9.0
vcrpy==1.2.0
websocket-client==0.25.0
wrapt==1.10.2
coverage==3.7.1
pyfiglet==0.7.4
tox==2.0.1
